requirements.txt
torch==2.6.0  # CPU-only; we'll add CUDA separately
torchvision==0.21.0
transformers==4.51.0
tokenizers==0.21.1
accelerate==1.6.0
bitsandbytes==0.44.1
huggingface-hub==0.30.1
safetensors==0.5.3
numpy==1.26.4
pandas==2.2.3
datasets==3.1.0
fastapi==0.115.9  # Updated to match chromadb
uvicorn==0.32.0
python-multipart==0.0.12
pillow==11.0.0
chromadb==1.0.0
httpx==0.27.2
langchain==0.3.7
openai==1.51.2
sentence-transformers==3.2.1